requirements.txt
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
pytest>=7.0.0,<9.0.0
jupyter>=1.0.0
google>=3.0.0
google-genai>=1.55.0
